emit success toast with hash on tx success
